Following benchmarking:
The file input remains unoptimized -- It trades speed for safety and 'sending'
one event at a time.

The ZMQ4 inputs and Redis output have been optimized to work on chunks of
events so as to spend less time in atomic memory. This has worked out quite
well. Happy to test in production now for some real numbers.